requirements.txt
Flask
mysql-connector-python
pandas
openpyxl
scikit-learn
gunicorn
wikipedia
google-generativeai
python-dotenv
twilio
qrcode
Pillow
pymongo
dnspython
flask-pymongo
